+++ View/volgendeFilm-succes.handlebars
@@ -5,13 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Forms</title>
     <link rel="stylesheet" href="style/style.css">
-<<<<<<< HEAD:views/volgendeFilm-succes.handlebars
-  </head>
-  <body>
-=======
 </head>
 <body>
->>>>>>> topic-mvc:View/volgendeFilm-succes.handlebars
     <h1>Oke {{ info.name }}, Which movie now?</h1>
     <form id="form" method="POST" action="/succes">
 
@@ -34,14 +29,7 @@
           <button>Submit</button>
       </label>
     </form>
-<<<<<<< HEAD:views/volgendeFilm-succes.handlebars
 
     <script src="server.js"></script>
   </body>
 </html>
-=======
-    
-    <script src="../server.js"></script>
-</body>
-</html>
->>>>>>> topic-mvc:View/volgendeFilm-succes.handlebars
